Add images inside icons folder because thats how we publish
