+++ src/icons/images/ExpediaImage.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { createIcon } from "../icons/helpers/icon";
+import { createIcon } from "../helpers/icon";
 
 export const ExpediaImage = createIcon((props) => {
     return (
