+++ src/icons/images/XolaLogo.jsx
