+++ src/icons/index.js
@@ -43,7 +43,7 @@ export { ChevronUpIcon } from "./ChevronUpIcon";
 export { ChipIcon } from "./ChipIcon";
 export { CircleCheckIcon } from "./CircleCheckIcon";
 export { CircleCrossIcon } from "./CircleCrossIcon";
-export { CircleCrossImage } from "../images/CircleCrossImage";
+export { CircleCrossImage } from "./images/CircleCrossImage";
 export { CircleCrownIcon } from "./CircleCrownIcon";
 export { CircleDollarIcon } from "./CircleDollarIcon";
 export { CircleDotIcon } from "./CircleDotIcon";
@@ -52,7 +52,7 @@ export { CircleInfoIcon } from "./CircleInfoIcon";
 export { CircleKeyIcon } from "./CircleKeyIcon";
 export { CirclePauseIcon } from "./CirclePauseIcon";
 export { CirclePlusIcon } from "./CirclePlusIcon";
-export { CirclePlusImage } from "../images/CirclePlusImage";
+export { CirclePlusImage } from "./images/CirclePlusImage";
 export { CircleSubtractIcon } from "./CircleSubtractIcon";
 export { ClipboardIcon } from "./ClipboardIcon";
 export { ClockAltIcon } from "./ClockAltIcon";
@@ -63,7 +63,7 @@ export { CollapseIcon } from "./CollapseIcon";
 export { CommentAltIcon } from "./CommentAltIcon";
 export { CommentIcon } from "./CommentIcon";
 export { CompassIcon } from "./CompassIcon";
-export { ComputerImage } from "../images/ComputerImage";
+export { ComputerImage } from "./images/ComputerImage";
 export { CopyIcon } from "./CopyIcon";
 export { CouponIcon } from "./CouponIcon";
 export { CrmIcon } from "./CrmIcon";
@@ -89,9 +89,9 @@ export { EmailCheckedIcon } from "./EmailCheckedIcon";
 export { EmailIcon } from "./EmailIcon";
 export { EmailResendIcon } from "./EmailResendIcon";
 export { EmptyChecklistIcon } from "./EmptyChecklistIcon";
-export { EmvImage } from "../images/EmvImage";
+export { EmvImage } from "./images/EmvImage";
 export { EquipmentIcon } from "./EquipmentIcon";
-export { ExpediaImage } from "../images/ExpediaImage";
+export { ExpediaImage } from "./images/ExpediaImage";
 export { ExportIcon } from "./ExportIcon";
 export { EyeIcon } from "./EyeIcon";
 export { FacebookIcon } from "./Logos/FacebookIcon";
@@ -193,7 +193,7 @@ export { UserAddIcon } from "./UserAddIcon";
 export { UserChangedIcon } from "./UserChangedIcon";
 export { UserIcon } from "./UserIcon";
 export { UserSubtractIcon } from "./UserSubtractIcon";
-export { VerifoneImage } from "../images/VerifoneImage";
+export { VerifoneImage } from "./images/VerifoneImage";
 export { VeteranIcon } from "./VeteranIcon";
 export { ViewNotesIcon } from "./ViewNotesIcon";
 export { VisaIcon } from "./CreditCards/VisaIcon";
