+++ src/icons/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@xola/icons",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "Xola's icon set",
   "main": "./index.js",
   "scripts": {
