+++ src/index.js
@@ -48,9 +48,6 @@ export { ImageUpload } from "./components/ImageUpload";
 export { Currency } from "./components/Utilities/Currency";
 export { Phone } from "./components/Utilities/Phone";
 export { Number } from "./components/Utilities/Number";
-export { XolaLogoCircle } from "./images/XolaLogoCircle";
-export { XolaLogo } from "./images/XolaLogo";
-export { XolaLogoSimple } from "./images/XolaLogoSimple";
 
 // Helpers
 export { almostZero, numberFormat, roundNumber } from "./helpers/numbers";
